% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-chisquare.R
\name{vdist_chisquare_plot}
\alias{vdist_chisquare_plot}
\title{Visualize chi square distribution}
\usage{
vdist_chisquare_plot(df = 3, normal = FALSE)
}
\arguments{
\item{df}{Degrees of freedom.}

\item{normal}{If \code{TRUE}, normal curve with same \code{mean} and
\code{sd} as the chi square distribution is drawn.}
}
\description{
Visualize how changes in degrees of freedom affect the shape of
the chi square distribution. Compute & visualize quantiles out of given
probability and probability from a given quantile.
}
\examples{
# visualize chi square distribution
vdist_chisquare_plot()
vdist_chisquare_plot(df = 5)
vdist_chisquare_plot(df = 5, normal = TRUE)

}
\seealso{
\code{\link[stats]{Chisquare}}
}
