% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-app.R
\name{vdist_launch_app}
\alias{vdist_launch_app}
\title{Launch shiny app}
\usage{
vdist_launch_app()
}
\description{
Launches shiny app for visualizing distributions.
}
\examples{
\dontrun{
vdist_launch_app ()
}
}
