% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-binomial.R
\name{vdist_binom_plot}
\alias{vdist_binom_plot}
\title{Visualize binomial distribution}
\usage{
vdist_binom_plot(n, p)
}
\arguments{
\item{n}{Number of trials.}

\item{p}{Aggregate probability.}
}
\description{
Visualize how changes in number of trials and the probability of
success affect the shape of the binomial distribution. Compute & visualize
probability from a given quantile and quantiles out of given probability.
}
\examples{
# visualize binomial distribution
vdist_binom_plot(10, 0.3)

}
\seealso{
\code{\link[stats]{Binomial}}
}
