% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-chisquare.R
\name{vdist_chisquare_plot}
\alias{vdist_chisquare_plot}
\alias{vdist_chisquare_perc}
\alias{vdist_chisquare_prob}
\title{Visualize chi square distribution}
\usage{
vdist_chisquare_plot(df = 3, normal = FALSE, print_plot = TRUE)

vdist_chisquare_perc(probs = 0.95, df = 3, type = c("lower",
  "upper"), print_plot = TRUE)

vdist_chisquare_prob(perc, df, type = c("lower", "upper"),
  print_plot = TRUE)
}
\arguments{
\item{df}{Degrees of freedom.}

\item{normal}{If \code{TRUE}, normal curve with same \code{mean} and
\code{sd} as the chi square distribution is drawn.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}

\item{probs}{Probability value.}

\item{type}{Lower tail or upper tail.}

\item{perc}{Quantile value.}
}
\description{
Visualize how changes in degrees of freedom affect the shape of
the chi square distribution. Compute & visualize quantiles out of given
probability and probability from a given quantile.
}
\examples{
# visualize chi square distribution
vdist_chisquare_plot()
vdist_chisquare_plot(df = 5)
vdist_chisquare_plot(df = 5, normal = TRUE)

# visualize quantiles out of given probability
vdist_chisquare_perc(0.165, 8, 'lower')
vdist_chisquare_perc(0.22, 13, 'upper')

# visualize probability from a given quantile.
vdist_chisquare_prob(13.58, 11, 'lower')
vdist_chisquare_prob(15.72, 13, 'upper')

}
\seealso{
\code{\link[stats]{Chisquare}}
}
