% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-f.R
\name{vdist_f_plot}
\alias{vdist_f_plot}
\title{Visualize f distribution}
\usage{
vdist_f_plot(num_df = 4, den_df = 30, normal = FALSE)
}
\arguments{
\item{num_df}{Degrees of freedom associated with the numerator of f statistic.}

\item{den_df}{Degrees of freedom associated with the denominator of f statistic.}

\item{normal}{If \code{TRUE}, normal curve with same \code{mean} and
\code{sd} as the F distribution is drawn.}
}
\description{
Visualize how changes in degrees of freedom affect the
shape of the F distribution. Compute & visualize quantiles out of given
probability and probability from a given quantile.
}
\examples{
# visualize F distribution
vdist_f_plot()
vdist_f_plot(6, 10, normal = TRUE)

}
\seealso{
\code{\link[stats]{FDist}}
}
