% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdist-normal.R
\name{vdist_normal_plot}
\alias{vdist_normal_plot}
\title{Visualize normal distribution}
\usage{
vdist_normal_plot(mean = 0, sd = 1)
}
\arguments{
\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}
}
\description{
Visualize how changes in mean and standard deviation affect the
shape of the normal distribution. Compute & visualize quantiles out of given
probability  and probability from a given quantile.
}
\examples{
# visualize normal distribution
vdist_normal_plot()
vdist_normal_plot(mean = 2, sd = 0.6)

}
\seealso{
\code{\link[stats]{Normal}}
}
